Move cache/converter.go to util/converter/converter.go
